# [PERF] models: Do not cache HTML fields on record creation